+++ src/Monolog/Handler/BrowserConsoleHandler.php
@@ -70,14 +70,19 @@ class BrowserConsoleHandler extends AbstractProcessingHandler
             } elseif ($format === 'js') {
                 static::writeOutput(static::generateScript());
             }
-            static::reset();
+            static::resetStatic();
         }
+    }
+
+    public function reset()
+    {
+        self::resetStatic();
     }
 
     /**
      * Forget all logged records
      */
-    public static function reset(): void
+    public static function resetStatic(): void
     {
         static::$records = [];
     }