+++ src/Monolog/Registry.php
@@ -130,4 +130,14 @@ class Registry
     {
         return self::getInstance($name);
     }
+
+    /**
+     * @internal
+     */
+    public function getClass($object)
+    {
+        $class = \get_class($object);
+
+        return 'c' === $class[0] && 0 === strpos($class, "class@anonymous\0") ? get_parent_class($class).'@anonymous' : $class;
+    }
 }